Move atom renderers, countries
