--- indigo_api/views/countries.py
@@ -1,9 +1,0 @@
-from rest_framework import viewsets, mixins
-from ..serializers import CountrySerializer
-
-from ..models import Country
-
-
-class CountryViewSet(mixins.ListModelMixin, viewsets.GenericViewSet):
-    queryset = Country.objects.prefetch_related('localities', 'country')
-    serializer_class = CountrySerializer
