+++ indigo_content_api/urls.py
@@ -1,9 +1,14 @@
-from django.conf.urls import url
+from django.conf.urls import url, include
+from rest_framework.routers import DefaultRouter
 
 import views
 
+router = DefaultRouter(trailing_slash=False)
+router.register(r'countries', views.CountryViewSet, base_name='country')
 
 urlpatterns = [
+    url(r'^', include(router.urls)),
+
     # --- public content API ---
     # viewing a specific document identified by FRBR URI fragment,
     # this requires at least 4 components in the FRBR URI,
